backend/app/requirements.txt
Flask
psycopg2-binary
requests
postgrest